pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "lv3-platform-cli"
version = "0.1.0"
description = "Unified operator CLI for the LV3 platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["PyYAML>=6.0"]

[project.scripts]
lv3 = "scripts.lv3_cli:main"

[tool.setuptools.packages.find]
include = ["platform*", "scripts*"]
